Add support for dot notation, fix some whitespace (#8)
* Add support for dot notation, fix some whitespace

* Update package.json

* Remove oldPassword from tests

Storing the password and resetting the object wasn't needed because Mocha gives us a clean slate for every test it performs

* Remove usernameField dot notation "support"

Lots of edge-cases that weren't being handled with this implementation. Better to remove these changes for now and come back to this later.

For now, if you need dot notation for your usernameField you're better off writing a custom verifier function.